pyproject.toml
[tool.poetry]
name = "stalk"
version = "0.1.0"
description = ""
authors = [
	"David <devid02ru@gmail.com>",
	"Maxim <sinelnikov-web@mail.ru>",
	"Alexey <tixigam@gmail.com>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
aiohttp = "^3.10.10"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.12.3"
bs4 = "^0.0.2"
fake-headers = "^1.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
